requirements.txt
black==22.1.0
certifi==2021.5.30
cfgv==3.3.1
chardet==4.0.0
click==8.0.4
colorama==0.4.4
distlib==0.3.4
filelock==3.6.0
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
identify==2.4.11
idna==2.10
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
msgpack==1.0.2
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.28
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.5.1
pre-commit==2.17.0
protobuf==3.19.4
psycopg2==2.9.3
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==6.0
requests==2.25.1
signalrcore==0.9.2
six==1.16.0
SQLAlchemy==1.4.32
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.5
virtualenv==20.13.3
websocket-client==0.54.0
Werkzeug==2.0.3
